+++ src/partials/hero.html
@@ -1,6 +1,6 @@
     <section class="hero section">
         <div class="container">
-            <div class="hero__keyboard keyboard">
+            <div class="hero__keyboard">
                 <h1 class="hero__title">A party without cocktails is not like a party</h1>
                 <picture>
                     <source srcset="./img/pictures/desktop/hero_501x580.png" media="(min-width:1280px)" />
@@ -15,10 +15,10 @@
                     <img class="hero__img" src="./img/pictures/desktop/hero_501x580.png" alt="Amaizing cocktail" />
                 </picture>
                 <p class="hero__text">Search your favorite cocktail by ABC</p>
-                <div class="keyboard__list">
-                    <ul class="first--line">
+                <div class="keyboard__list keyboard">
+                    <ul class="first--line ">
                         <li>
-                            <div class="keyboard--item">A</div>
+                            <div class="keyboard--item ">A</div>
                         </li>
                         <li>
                             <div class="keyboard--item">B</div>
