+++ src/sass/layout/_hero.scss
@@ -1,7 +1,5 @@
 .hero {
-    position: relative;
     width: 100%;
-    background-color: #FCFCFC;
     padding-top: 30px;
     padding-bottom: 20px;
     margin-left: auto;
@@ -20,13 +18,19 @@
     }
 
 }
+.hero__keyboard{
+    position: relative;
+    width: 100%;
+    margin-left: auto;
+    margin-right: auto;
+}
 
 .hero__img {
     margin-bottom: 28px;
 
     @media screen and (min-width: 768px) {
         position: absolute;
-        top: 38px;
+        top: -64px;
         right: 32px;
         z-index: 5;
     }
@@ -96,6 +100,7 @@
 }
 
 .keyboard__list{
+ 
     @media screen and (min-width: 768px) {
         display: inline-block;
     }
@@ -115,6 +120,7 @@
     text-align: center;
     
     color: #5F6775;
+    cursor: pointer;
     
     }
     
@@ -147,4 +153,3 @@
     justify-content: center;
     gap: 1px;
 }
-
